+++ src/client/app/widgets/plots/frame.js
@@ -1,8 +1,7 @@
 var Widget = require('../common/widget'),
     html = require('nanohtml'),
     {urlParser} = require('../utils'),
-    locales = require('../../locales'),
-    localUrlRe = /(^127\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}$)|(^10\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}$)|(^172\.1[6-9]{1}[0-9]{0,1}\.[0-9]{1,3}\.[0-9]{1,3}$)|(^172\.2[0-9]{1}[0-9]{0,1}\.[0-9]{1,3}\.[0-9]{1,3}$)|(^172\.3[0-1]{1}[0-9]{0,1}\.[0-9]{1,3}\.[0-9]{1,3}$)|(^192\.168\.[0-9]{1,3}\.[0-9]{1,3}$)/
+    locales = require('../../locales')
 
 class Frame extends Widget {
 
@@ -54,7 +53,7 @@ class Frame extends Widget {
 
         var parser = urlParser(this.value)
 
-        if (this.value && parser.hostname.match(localUrlRe)) {
+        if (this.value && parser.isLocal()) {
             this.frame.setAttribute('src', this.value)
             this.widget.classList.remove('error')
             if (this.errorTextMounted) {